Julia modtest works

+++ fcode/hamiltonian/jmodtest.cpp
@@ -7,7 +7,7 @@
 #include "../config/load/cpp_config.hpp"
 #include "../hamiltonian/band_structure.hpp"
 
-extern "C" float epsilon_julia(int n, double *k, int size) {
+extern "C" double epsilon_julia(int n, double *k, int size) {
     Vec kvec;
     for (int i = 0; i < size; i++) {
         kvec(i) = k[i];
